# api-rest/src/main/java/com/sitionix/wagssox/api/dto/ErrorDTO.java
package com.sitionix.wagssox.api.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ErrorDTO {

    private Integer code;
    private String title;
    private String details;
}