# api-rest/src/main/resources/openapi-rest.yml
openapi: 3.0.0
info:
  title: Workspace Aggregation Service API
  version: 0.0.1
paths:
  /api/v1/sites:
    get:
      tags:
        - Site
      operationId: getSites
      summary: Get workspace sites
      description: Returns paginated workspace site cards for a user.
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: size
          required: true
          schema:
            type: integer
            enum: [20]
      responses:
        "200":
          description: Sites page response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceSitesResponseDTO"
        "400":
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
components:
  schemas:
    WorkspaceSitesResponseDTO:
      type: object
      required:
        - items
        - page
        - size
        - hasNext
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceSiteCardResponseDTO"
        page:
          type: integer
        size:
          type: integer
          enum: [20]
        hasNext:
          type: boolean
    WorkspaceSiteCardResponseDTO:
      type: object
      required:
        - siteId
        - name
        - status
        - createdAt
        - updatedAt
      properties:
        siteId:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        type:
          nullable: true
          type: string
          enum: [PORTFOLIO, BUSINESS, BLOG, STORE, LANDING, OTHER]
        description:
          nullable: true
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorDTO:
      type: object
      required:
        - code
        - title
        - details
      properties:
        code:
          type: integer
        title:
          type: string
        details:
          type: string